extract/requirements.txt
requests==2.32.3
pandas==2.2.3
snowflake-connector-python[pandas]==3.12.3
python-dotenv==1.0.1
firecrawl-py>=1.0.0
pytest>=7.4
